Reorganiza la estructura de archivos y paquetes
Se movieron las clases que:
- Representan tipos de datos a "datos".
- Representan datos de empleados a "datos.empleado".
- Se encargan de la entrada-salida de datos a "entradasalida".
- Se encargan del manejo de base de datos a "manejadorbd".
- Permiten al usuario interactuar con la lógica del programa a "ui".

Se crearon y eliminaron otras clases.
