+++ nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Ailu\\AppData\\Roaming\\NetBeans\\25\\build.properties
+user.properties.file=C:\\Users\\user\\AppData\\Roaming\\NetBeans\\20\\build.properties
