--- src/gestormundomarino/Capacidad.java
@@ -1,7 +1,0 @@
-package gestormundomarino;
-
-public interface Capacidad {
-
-    public boolean login(String usuario, String contrasenia);
-    
-}
