--- src/gestormundomarino/Capacidades.java
@@ -1,9 +1,0 @@
-package gestormundomarino;
-
-
-public class Capacidades {
-
-    public Capacidades() {
-    }
-    
-}
