+++ src/gestormundomarino/datos/ACargoDe.java
@@ -1,4 +1,4 @@
-package gestormundomarino;
+package gestormundomarino.datos;
 
 
 public class ACargoDe extends Dato {
