+++ src/gestormundomarino/datos/AtiendeA.java
@@ -1,4 +1,4 @@
-package gestormundomarino;
+package gestormundomarino.datos;
 
 //Cuidador AtiendeA Especie
