+++ src/gestormundomarino/datos/Recorre.java
@@ -1,4 +1,4 @@
-package gestormundomarino;
+package gestormundomarino.datos;
 
 //Itinerario Recorre Zona
 public class Recorre extends Dato {
