+++ src/gestormundomarino/ui/MenuCuidador.java
@@ -1,4 +1,4 @@
-package gestormundomarino;
+package gestormundomarino.ui;
 
 import java.util.List;
